add route guards to edit and delete

+++ JS BackEnd/earth-treasure-vaul/src/controller/stoneController.js
@@ -38,11 +38,16 @@ router.get('/:stoneId/details', async (req, res) => {
 
 router.get('/:stoneId/delete', async (req, res) =>{
     const stoneId = req.params.stoneId;
+    const isCreator = stone.owner._id == req.user?._id;
 
-    try{
-        await stoneService.deleteStone(stoneId)
-    }catch(err){
-        res.render('stones/dashboard', {error: getErrorMessage(err)})
+    if(!isCreator){
+        res.redirect('/stones/dashboard')
+    }else{
+         try{
+            await stoneService.deleteStone(stoneId)
+        }catch(err){
+            res.render('stones/dashboard', {error: getErrorMessage(err)})
+        }
     }
 })
 
@@ -50,21 +55,27 @@ router.get('/:stoneId/edit', async (req, res) => {
     const stoneId = req.params.stoneId;
     const stone = await stoneService.findOne(stoneId)
 
-    res.render('stones/edit', { stone })
+    const isCreator = stone.owner._id == req.user?._id;
+
+    if(!isCreator){
+        res.redirect('/stones/dashboard')
+    }else{
+        res.render('stones/edit', { stone })
+    }
 })
 
 router.post('/:stoneId/edit', async (req, res) => {
     const stoneId = req.params.stoneId;
     const stoneData = req.body;
-
+    
     const stone = await stoneService.findOne(stoneId)
+
     try{
         await stoneService.editStone(stone, stoneData)
         res.redirect(`/stones/${stoneId}/details`)
     }catch(err){
         res.render(`stones/${stoneId}/edit`, { stone, error: getErrorMessage(err) })
     }
-
 })
 
 router.get('/search', (req, res) => {
